requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-dotenv
cryptography>=41.0.0
pydantic[email]
python-multipart
pydantic
bcrypt
python-jose
aiofiles
cloudinary
authlib
httpx
itsdangerous
alembic